+++ scripts/create_resources/test_resources.sh
@@ -6,15 +6,10 @@ REPO_ROOT=$(git rev-parse --show-toplevel)
 # ensure that the command below is run from the root of the repository
 cd "$REPO_ROOT"
 
-# # remove this when you have implemented the script
-# echo "TODO: replace the commands in this script with the sequence of components that you need to run to generate test_resources."
-# echo "  Inside this script, you will need to place commands to generate example files for each of the 'src/api/file_*.yaml' files."
-# exit 1
-
 set -e
 
 RAW_DATA=resources_test/common
-DATASET_DIR=resources_test/denoising
+DATASET_DIR=resources_test/task_denoising
 
 mkdir -p $DATASET_DIR
 
@@ -22,31 +17,29 @@ mkdir -p $DATASET_DIR
 viash run src/data_processors/process_dataset/config.vsh.yaml -- \
   --input $RAW_DATA/cxg_mouse_pancreas_atlas/dataset.h5ad \
   --output_train $DATASET_DIR/cxg_mouse_pancreas_atlas/train.h5ad \
-  --output_test $DATASET_DIR/cxg_mouse_pancreas_atlas/test.h5ad \
-  --output_solution $DATASET_DIR/cxg_mouse_pancreas_atlas/solution.h5ad
+  --output_test $DATASET_DIR/cxg_mouse_pancreas_atlas/test.h5ad
 
 # run one method
 viash run src/methods/magic/config.vsh.yaml -- \
-    --input_train $DATASET_DIR/pancreas/train.h5ad \
-    --output $DATASET_DIR/pancreas/denoised.h5ad
+    --input_train $DATASET_DIR/cxg_mouse_pancreas_atlas/train.h5ad \
+    --output $DATASET_DIR/cxg_mouse_pancreas_atlas/denoised.h5ad
 
 # run one metric
 viash run src/metrics/poisson/config.vsh.yaml -- \
-    --input_denoised $DATASET_DIR/pancreas/denoised.h5ad \
-    --input_test $DATASET_DIR/pancreas/test.h5ad \
-    --output $DATASET_DIR/pancreas/score.h5ad
+    --input_prediction $DATASET_DIR/cxg_mouse_pancreas_atlas/denoised.h5ad \
+    --input_test $DATASET_DIR/cxg_mouse_pancreas_atlas/test.h5ad \
+    --output $DATASET_DIR/cxg_mouse_pancreas_atlas/score.h5ad
 
 # write manual state.yaml. this is not actually necessary but you never know it might be useful
 cat > $DATASET_DIR/cxg_mouse_pancreas_atlas/state.yaml << HERE
 id: cxg_mouse_pancreas_atlas
 train: !file train.h5ad
 test: !file test.h5ad
-solution: !file solution.h5ad
 prediction: !file denoised.h5ad
 score: !file score.h5ad
 HERE
 
 # only run this if you have access to the openproblems-data bucket
-# aws s3 sync --profile op \
-#   "$DATASET_DIR" s3://openproblems-data/resources_test/denoising \
-#   --delete --dryrun
+aws s3 sync --profile OP \
+  "$DATASET_DIR" s3://openproblems-data/resources_test/task_denoising \
+  --delete --dryrun