+++ src/data_processors/process_dataset/script.py
@@ -5,7 +5,7 @@ import numpy as np
 
 ## VIASH START
 par = {
-    'input': "resources_test/common/pancreas/dataset.h5ad",
+    'input': "resources_test/common/cxg_mouse_pancreas_atlas/dataset.h5ad",
     'output_train': "train.h5ad",
     'output_test': "test.h5ad",
     'train_frac': 0.9,