+++ src/methods/knn_smoothing/script.py
@@ -3,7 +3,7 @@ import anndata as ad
 
 ## VIASH START
 par = {
-    'input_train': 'resources_test/denoising/pancreas/train.h5ad',
+    'input_train': 'resources_test/task_denoising/cxg_mouse_pancreas_atlas/train.h5ad',
     'output': 'output_knn.h5ad',
 }
 meta = {