+++ src/methods/saver/script.R
@@ -5,7 +5,7 @@ library(Matrix, warn.conflicts = FALSE)
 
 ## VIASH START
 par <- list(
-  input_train = "resources_test/denoising/pancreas/train.h5ad",
+  input_train = "resources_test/task_denoising/cxg_mouse_pancreas_atlas/train.h5ad",
   norm = "log",
   output = "output.h5ad"
 )